requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
Django==5.2.1
django-telegram-login==0.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pillow==11.2.1
python-telegram-bot>=20.3
sniffio==1.3.1
sqlparse==0.5.3
python-dotenv>=1.0.0
